Large clean up.
- Finished all documentation.
- Cleaned up includes and forward declarations.
- Fixed all warnings.

+++ BitBrain/Source/ImageProcessing/SymbolExtractor.cpp
@@ -5,7 +5,6 @@
 #include "../Utility/Math/Functions.h"
 
 #include <queue>
-#include <iostream>
 
 namespace BB
 {
@@ -59,7 +58,7 @@ namespace BB
 
 		float scale = Math::Min(22 / (float)tempImage.Width(), 22 / (float)tempImage.Height());
 
-		tempImage = Image::Resize(tempImage, Vector2I(tempImage.Width() * scale, tempImage.Height() * scale));
+		tempImage = Image::Resize(tempImage, Vector2I(tempImage.Width() * (int)scale, tempImage.Height() * (int)scale));
 
 		Image finalImage(28, 28);
